+++ docs/.vitepress/vitepress/styles/base.scss
@@ -76,6 +76,7 @@ h6:focus .header-anchor {
 
 h1 {
   margin-top: 1.5rem;
+  margin-bottom: 2rem;
   font-size: 1.9rem;
 }
 
@@ -127,6 +128,7 @@ textarea {
 }
 
 a {
+  font-weight: 500;
   text-decoration: inherit;
   color: var(--brand-color);
 
@@ -204,7 +206,7 @@ li > ol {
 
 blockquote {
   margin: 1rem 0;
-  border-left: 0.2rem solid var(--border-color-light);
+  border-left: 0.2rem solid var(--el-border-color);
   padding: 0.25rem 0 0.25rem 1rem;
   font-size: 1rem;
   color: var(--text-color-lighter);
