+++ docs/.vitepress/vitepress/styles/content.scss
@@ -9,11 +9,12 @@
     &.has-sidebar {
       padding-left: var(--sidebar-width-sm);
       @include respond-to('xlg') {
-        padding-left: var(--sidebar-width-md);
+        padding-left: var(--vp-sidebar-width-small);
       }
       @include respond-to('xxl') {
         padding-left: calc(
-          (100vw - var(--app-max-width)) / 2 + var(--sidebar-width-md)
+          (100vw - var(--vp-screen-max-width)) / 2 +
+            var(--vp-sidebar-width-small)
         );
       }
     }
