+++ docs/.vitepress/vitepress/styles/content/doc-content.scss
@@ -2,11 +2,24 @@
 @use './table-of-content';
 
 .doc-content-wrapper {
-  --toc-width: 144px;
+  --vp-content-width: 800px;
+
   padding: 32px 24px 96px;
+
   .doc-content-container {
-    max-width: 60rem;
-    margin: 0 auto;
+    max-width: var(--vp-content-width);
+
+    .doc-content {
+      width: 100%;
+    }
+
+    @include respond-to('xlg') {
+      width: var(--vp-content-width);
+    }
+
+    @include respond-to('xxl') {
+      width: var(--vp-content-width);
+    }
   }
 
   @include respond-to('md') {
@@ -18,9 +31,5 @@
   @include respond-to('xlg') {
     padding: 64px 0 96px 64px;
     display: flex;
-    .doc-content-container {
-      min-width: 38.75rem;
-      margin: 0;
-    }
   }
 }
