+++ docs/.vitepress/vitepress/styles/content/table-of-content.scss
@@ -3,13 +3,15 @@
 .toc-wrapper {
   display: none;
   padding-left: 64px;
+
   .toc-content {
     position: sticky;
     top: calc(var(--header-height) + 32px);
     margin-top: 32px;
     padding: 4px 8px 4px 12px;
     margin-bottom: 32px;
-    width: var(--toc-width);
+    width: 200px;
+
     .toc-marker {
       opacity: 0;
       position: absolute;
