+++ docs/.vitepress/vitepress/styles/navbar.scss
@@ -24,7 +24,7 @@
     display: flex;
     justify-content: space-between;
     margin: 0 auto;
-    max-width: var(--app-max-width);
+    max-width: var(--vp-screen-max-width);
 
     .content {
       display: flex;
