+++ docs/.vitepress/vitepress/styles/sidebar.scss
@@ -64,17 +64,20 @@
   }
 
   @include respond-to('md') {
-    width: var(--sidebar-width-md);
+    width: var(--vp-sidebar-width-small);
   }
 
   @include respond-to('lg') {
     top: var(--header-height);
     transform: translate(0);
     border-right: 1px solid var(--border-color);
+    width: var(--vp-sidebar-width-small);
   }
 
   @include respond-to('xxl') {
-    padding: 24px 32px 96px calc((100% - var(--app-max-width)) / 2);
-    width: calc((100% - var(--app-max-width)) / 2 + var(--sidebar-width-md));
+    padding: 0 32px 96px calc((100% - var(--vp-screen-max-width)) / 2.2);
+    width: calc(
+      (100% - var(--vp-screen-max-width)) / 2 + var(--vp-sidebar-width-small)
+    );
   }
 }
