+++ docs/en-US/guide/theming.md
@@ -7,8 +7,9 @@ lang: en-US
 
 Element Plus uses BEM-styled CSS so that you can override styles easily. But if
 you need to replace styles at a large scale, e.g. change the theme color from
-blue to orange or green, maybe overriding them one by one is not a good idea. We
-provide four ways to change the style variables.
+blue to orange or green, maybe overriding them one by one is not a good idea.
+
+We provide four ways to change the style variables.
 
 ## Change theme color
 
@@ -24,7 +25,7 @@ You can find SCSS variables in [`packages/theme-chalk/src/common/var.scss`](http
 
 ::: warning
 
-We use sass modules ([sass:map](https://sass-lang.com/documentation/values/maps)...) to refactor all SCSS variables.
+We use sass modules ([sass:map](https://sass-lang.com/documentation/values/maps)...) and `@use` to refactor all SCSS variables.
 
 > [Introducing Sass Modules | CSS-TRICKS](https://css-tricks.com/introducing-sass-modules/)
 
@@ -68,6 +69,16 @@ $colors: map.deep-merge(
 ### How to override it?
 
 If your project also uses SCSS, you can directly change Element Plus style variables. Create a new style file, e.g. `styles/element/index.scss`:
+
+::: warning
+
+You should use `@use 'xxx.scss' as *;` instead of `@import 'xxx.scss';`.
+
+Because the sass team said they will remove it eventually.
+
+> [Sass: @use](https://sass-lang.com/documentation/at-rules/use) vs [Sass: @import](https://sass-lang.com/documentation/at-rules/import)
+
+:::
 
 ```scss
 // styles/element/index.scss
